# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mavros Connect API - SaaS Hotspot Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"

    post:
      operationId: createVoucher
      tags: [vouchers]
      summary: Generate a single voucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherInput"
      responses:
        "201":
          description: Voucher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
  /vouchers/bulk:
    post:
      operationId: bulkCreateVouchers
      tags: [vouchers]
      summary: Generate multiple vouchers in bulk
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkVoucherInput"
      responses:
        "201":
          description: Vouchers generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
  /vouchers/{id}:
    get:
      operationId: getVoucher
      tags: [vouchers]
      summary: Get voucher by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"

    delete:
      operationId: deleteVoucher
      tags: [vouchers]
      summary: Delete / revoke a voucher
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /vouchers/validate:
    post:
      operationId: validateVoucher
      tags: [vouchers]
      summary: Validate a voucher code (captive portal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherValidateInput"
      responses:
        "200":
          description: Voucher validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherValidationResult"

  # ─── PAYMENTS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Package:
      type: object
      required: [id, tenantId, name, price, duration, durationUnit, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        duration:
          type: integer
        durationUnit:
          type: string
          enum: [minutes, hours, days, weeks, months]
        downloadSpeed:
          type: ["integer", "null"]
        uploadSpeed:
          type: ["integer", "null"]
        dataCapMb:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    Voucher:
      type: object
      required: [id, tenantId, packageId, code, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        packageId:
          type: integer
        locationId:
          type: ["integer", "null"]
        code:
          type: string
        status:
          type: string
          enum: [unused, active, used, expired, revoked]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        usedAt:
          type: ["string", "null"]
          format: date-time
        usedByMac:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        package:
          $ref: "#/components/schemas/Package"
    VoucherInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: integer
        locationId:
          type: integer
        expiresAt:
          type: string
          format: date-time
    BulkVoucherInput:
      type: object
      required: [packageId, quantity]
      properties:
        packageId:
          type: integer
        locationId:
          type: integer
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
        expiresAt:
          type: string
          format: date-time
    VoucherValidateInput:
      type: object
      required: [code, tenantSlug]
      properties:
        code:
          type: string
        tenantSlug:
          type: string
        macAddress:
          type: string
    VoucherValidationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        message:
          type: string
        voucher:
          $ref: "#/components/schemas/Voucher"

    # ── Payment ──────────────────────────────────────────────────────────────
...